external_sync/pyproject.toml
[tool.poetry]
name = "keycloak-ekklesia"
version = "0.1.0"
description = ""
authors = ["Tobias dpausp <dpausp@posteo.de>"]

[tool.poetry.dependencies]
python = "^3.8"
typer = "^0.3.2"
python-keycloak = "^0.22.0"
eliot = "^1.12.0"
eliot-tree = "^19.0.1"
dataclasses-json = "^0.5.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
